@@ -1,5 +1,5 @@
 {
-    "name": "furfi",
+    "name": "svc-dex",
     "version": "0.1.0",
     "private": true,
     "dependencies": {
@@ -15,39 +15,56 @@
         "@ethersproject/providers": "^5.7.2",
         "@ethersproject/strings": "^5.7.0",
         "@ethersproject/units": "^5.7.0",
+        "@gelatonetwork/limit-orders-lib": "^4.5.0",
         "@mui/icons-material": "^5.10.9",
         "@mui/material": "^5.10.12",
         "@mui/styled-engine-sc": "^5.10.6",
         "@mui/styles": "^5.10.10",
         "@reduxjs/toolkit": "^1.8.6",
+        "@sentry/nextjs": "^7.36.0",
         "@tabler/icons": "^1.105.0",
         "@web3-react/core": "^6.1.9",
         "@web3-react/injected-connector": "^6.0.7",
         "@web3-react/walletconnect-connector": "^6.2.13",
         "@web3-react/walletlink-connector": "^6.2.14",
         "axios": "^1.2.1",
+        "big.js": "^6.2.1",
         "clsx": "^1.2.1",
+        "date-fns": "^2.29.3",
+        "decimal.js-light": "^2.5.1",
         "ethereum-multicall": "^2.15.0",
         "ethers": "^5.7.2",
         "from-exponential": "^1.1.1",
+        "i18next": "^22.4.9",
+        "i18next-browser-languagedetector": "^7.0.1",
+        "jotai": "^2.0.0",
+        "jsbi": "^4.3.0",
+        "localforage": "^1.10.0",
         "lodash": "^4.17.21",
         "moment": "^2.29.4",
-        "react": "^17.0.2",
+        "next": "^13.0.7",
+        "numeral": "^2.0.6",
+        "react": "^18.2.0",
         "react-circular-progressbar": "^2.1.0",
         "react-copy-to-clipboard": "^5.1.0",
         "react-device-detect": "^2.2.2",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
+        "react-i18next": "^12.1.4",
         "react-perfect-scrollbar": "^1.5.8",
         "react-qr-code": "^2.0.8",
         "react-redux": "^8.0.4",
         "react-router-dom": "^6.4.1",
         "react-scripts": "^4.0.3",
-        "react-slick": "^0.29.0",
         "react-toastify": "^9.0.3",
+        "react-transition-group": "^4.4.5",
+        "react-window": "^1.8.8",
         "redux-localstorage-simple": "^2.5.1",
-        "slick-carousel": "^1.8.1",
+        "redux-persist": "^6.0.0",
         "styled-components": "^5.3.6",
-        "swr": "^1.3.0",
+        "swr": "^2.0.1",
+        "tiny-invariant": "^1.3.1",
+        "tiny-warning": "^1.0.3",
+        "toformat": "^2.0.0",
         "typescript": "^4.6.3",
         "wagmi": "^0.6.7",
         "web-vitals": "^2.1.0",
@@ -79,8 +96,8 @@
     "devDependencies": {
         "@babel/core": "^7.19.6",
         "@babel/helper-module-transforms": "^7.19.6",
-        "@types/react": "^17.0.2",
-        "@types/react-dom": "^17.0.2",
+        "@types/react": "^18.0.17",
+        "@types/react-dom": "^18.0.6",
         "@types/react-redux": "^7.1.24",
         "@types/react-router-dom": "^5.3.3",
         "@typescript-eslint/eslint-plugin": "^5.42.0",
@@ -92,6 +109,7 @@
         "eslint-plugin-babel": "^5.3.1",
         "eslint-plugin-import": "^2.22.1",
         "eslint-plugin-jsx-a11y": "^6.4.1",
+        "eslint-plugin-lodash": "^7.4.0",
         "eslint-plugin-prettier": "^3.1.4",
         "eslint-plugin-react": "^7.22.0",
         "eslint-plugin-react-hooks": "^4.2.0",
