@@ -1,8 +1,6 @@
 import React from 'react'
 import { makeStyles } from '@mui/styles'
 import { Box, Typography } from '@mui/material'
-
-
 import TwitterIcon from '../../asset/images/twitter.svg'
 import TelegramIcon from '../../asset/images/telegram.svg'
 import DiscordIcon from '../../asset/images/discord.svg'
@@ -10,6 +8,7 @@ import MediumIcon from '../../asset/images/medium.svg'
 import FacebookIcon from '../../asset/images/facebook.svg'
 import InstagramIcon from '../../asset/images/instagram.svg'
 import SevenChainIcon from '../../asset/images/seven_chain_logo.png'
+import { useTranslation } from 'src/context/Localization'
 
 
 const useStyles = makeStyles(() => ({
@@ -17,8 +16,9 @@ const useStyles = makeStyles(() => ({
         padding: '20px 0',
         display: 'flex',
         justifyContent: 'center',
+        backgroundColor: 'rgb(245, 251, 255)',
         '& .MuiTypography-root': {
-            color: '#333'
+            color: '#555'
         },
         '& .MuiButton-root': {
             color: '#333',
@@ -31,23 +31,37 @@ const useStyles = makeStyles(() => ({
 
 function Footer() {
     const classes = useStyles()
+    const { t } = useTranslation()
     return (
         <div className={classes.footerView}>
-            <Box sx={{ display: 'flex', flexWrap: 'wrap' }}>
+            <Box sx={{ width: '100vw', display: 'flex', justifyContent: 'space-around', alignItems: 'center', flexWrap: 'wrap' }}>
                 <Box sx={{
                     display: 'flex',
-                    alignItems: 'center',
                     flexDirection: 'column',
                     m: 3
                 }}>
-                    <Typography sx={{ fontSize: '20px', mt: 2 }}>Social Links</Typography>
+                    <Box sx={{
+                        display: 'flex',
+                        flexDirection: 'column',
+                        justifyContent: 'center',
+                        alignItems: 'center',
+                        '& img': {
+                            width: '64px',
+                            height: '64px',
+                            marginBottom: '16px'
+                        }
+                    }}>
+                        <img src={SevenChainIcon} alt='svc_chain' />
+                        <Typography>{t('Copyright')}</Typography>
+                    </Box>
                     <Box sx={{
                         display: 'flex',
                         justifyContent: 'center',
                         '& img': {
                             width: 32,
                             height: 32,
-                            m: 1
+                            mx: 1,
+                            mt: 2
                         }
                     }}>
                         <img src={TwitterIcon} alt='twitter' />
@@ -58,41 +72,13 @@ function Footer() {
                         <img src={InstagramIcon} alt='instagram' />
                     </Box>
                 </Box>
-                <Box sx={{ my: 3, mx: 4 }}>
-                    <Typography sx={{ fontSize: '20px', mt: 2 }}>Useful Link</Typography>
-                    <Box sx={{ display: 'flex' }}>
-                        <Box sx={{ mr: 2, '& .MuiTypography-root': { pt: '7px', fontSize: '16px' } }}>
-                            <Typography>Seven Project</Typography>
-                            <Typography>Docs</Typography>
-                        </Box>
-                    </Box>
+                <Box sx={{ maxWidth: '840px', p: 2 }}>
+                    <Typography sx={{ fontSize: '24px', color: '#ffae5a !important', mb: 1.5 }}>{t('Disclaimer')}</Typography>
+                    <Typography sx={{ textIndent: '20px' }}>
+                        {t('Disclaimer Content')}
+                    </Typography>
                 </Box>
-                <Box sx={{ my: 2.7, mx: 4 }}>
-                    <Typography sx={{ fontSize: '20px', mt: 2 }}>Exchanges</Typography>
-                    <Box sx={{ display: 'flex' }}>
-                        <Box sx={{ mr: 2, '& .MuiTypography-root': { pt: '7px', fontSize: '16px' } }}>
-                            <Typography>Uniswap</Typography>
-                            <Typography>Quickswap</Typography>
-                            <Typography>Coinsbit</Typography>
-                        </Box>
-                    </Box>
-                </Box>
-                <Box sx={{
-                    my: 2.7,
-                    mx: 4,
-                    display: 'flex',
-                    flexDirection: 'column',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    '& img': {
-                        width: '64px',
-                        height: '64px',
-                        marginBottom: '16px'
-                    }
-                }}>
-                    <img src={SevenChainIcon} alt='svc_chain' />
-                    <Typography>Copyright © 2023.<br />By Seven Project Team</Typography>
-                </Box>
+
             </Box>
         </div>
     )
