@@ -1,6 +1,7 @@
 import { useCallback } from 'react'
 import { ConnectorNames } from 'src/config'
-import { DEFAULT_CHAIN_ID } from 'src/config/chains'
+import { DEFAULT_CHAIN_ID } from 'src/config/constants/chains'
+import { useTranslation } from 'src/context/Localization'
 import {
     ConnectorNotFoundError,
     SwitchChainError,
@@ -9,14 +10,17 @@ import {
     useDisconnect,
     useNetwork,
 } from 'wagmi'
-
+import useToast from './useToast'
 
 const useAuth = () => {
 
     const { connectAsync, connectors, isLoading } = useConnect()
     const { chain } = useNetwork()
     const { disconnectAsync } = useDisconnect()
-    const chainId = DEFAULT_CHAIN_ID;
+    const chainId = DEFAULT_CHAIN_ID
+
+    const { t } = useTranslation()
+    const { toastError } = useToast()
 
     const login = useCallback(
         async (connectorId: ConnectorNames) => {
@@ -26,9 +30,11 @@ const useAuth = () => {
                 return connected
             } catch (error) {
                 if (error instanceof ConnectorNotFoundError) {
+                    toastError(t('Provider Error'), t('No provider was found'))
                     console.log('wallet connector not found')
                 }
                 if (error instanceof SwitchChainError || error instanceof SwitchChainNotSupportedError) {
+                    toastError(t('Switch Network Error'), t('Unable to switch network, please try it on your wallet'))
                     console.log('Unable to switch network, please try it on your wallet')
                 }
             }
