@@ -1,32 +1,50 @@
-
+import useActiveWeb3React from "./useActiveWeb3React"
 import { useMemo } from 'react'
-import { getContract } from 'src/helper'
-import { erc20ABI, useSigner } from 'wagmi'
-import { useWeb3 } from './useWeb3'
-import StandardStrategyABI from '../config/abi/strategyStandard.json'
-import FurfiStrategyABI from '../config/abi/strategyFurfi.json'
-import StableStrategyABI from '../config/abi/strategyStable.json'
-import FreezeABI from '../config/abi/freezer.json'
-import ReferralABI from '../config/abi/referrer.json'
-import StakingABI from '../config/abi/staking.json'
 
-import addresses from 'src/config/address'
-import { DEFAULT_CHAIN_ID } from 'src/config/chains'
+import {
+    Erc20,
+    Erc20Bytes32,
+    Multicall,
+    Weth,
+} from 'src/config/abi/types'
 
-export enum Strategy {
-    Standard = 'standardStrategy',
-    Stable = 'stableStrategy',
-    Furfi = 'furfiStrategy'
+import { useProviderOrSigner } from './useProviderOrSigner'
+import { getMulticallAddress } from 'src/utils/addressHelper'
+import {
+    getErc20Contract,
+} from 'src/utils/contractHelper'
+
+// Imports below migrated from Exchange useContract.ts
+import { Contract } from '@ethersproject/contracts'
+import { WNATIVE } from "src/utils/token"
+import { ERC20_BYTES32_ABI } from 'src/config/abi/erc20'
+import ERC20_ABI from 'src/config/abi/erc20.json'
+import IPairABI from 'src/config/abi/IPair.json'
+import multiCallAbi from 'src/config/abi/multicall.json'
+import WETH_ABI from 'src/config/abi/weth.json'
+import { getContract } from 'src/utils'
+
+import { IPair } from 'src/config/abi/types'
+import { useActiveChainId } from './useActiveChainId'
+
+
+export const useERC20 = (address: string, withSignerIfPossible = true) => {
+    const providerOrSigner = useProviderOrSigner(withSignerIfPossible)
+    return useMemo(() => getErc20Contract(address, providerOrSigner), [address, providerOrSigner])
 }
 
-export function useContract(
-    address: string,
+
+// Code below migrated from Exchange useContract.ts
+
+// returns null on errors
+export function useContract<T extends Contract = Contract>(
+    address: string | undefined,
     ABI: any,
-    withSignerIfPossible = true
-) {
-    const { provider } = useWeb3()
-    const { data: signer } = useSigner()
-    const providerOrSigner = withSignerIfPossible ? signer ?? provider : provider
+    withSignerIfPossible = true,
+): T | null {
+    const { provider } = useActiveWeb3React()
+
+    const providerOrSigner = useProviderOrSigner(withSignerIfPossible) ?? provider
 
     const canReturnContract = useMemo(() => address && ABI && providerOrSigner, [address, ABI, providerOrSigner])
 
@@ -34,48 +52,33 @@ export function useContract(
         if (!canReturnContract) return null
         try {
             return getContract(address, ABI, providerOrSigner)
-        } catch (err) {
-            console.error('Failed to get contract', err)
+        } catch (error) {
+            console.error('Failed to get contract', error)
             return null
         }
-    }, [address, ABI, providerOrSigner])
+    }, [address, ABI, providerOrSigner, canReturnContract]) as T
 }
 
-export function useTokenContract(tokenAddress: string, withSignerIfPossible?: boolean) {
-    return useContract(tokenAddress, erc20ABI, withSignerIfPossible)
+export function useTokenContract(tokenAddress?: string, withSignerIfPossible?: boolean) {
+    return useContract<Erc20>(tokenAddress, ERC20_ABI, withSignerIfPossible)
 }
 
-
-export function useStrategyContract(pair: string, strategy: Strategy) {
-    let addr, abi
-    if (strategy === Strategy.Stable) {
-        addr = addresses['stableStrategy'][pair][DEFAULT_CHAIN_ID]
-        abi = StableStrategyABI
-    } else if (strategy === Strategy.Standard) {
-        addr = addresses['standardStrategy'][pair][DEFAULT_CHAIN_ID]
-        abi = StandardStrategyABI
-    } else if (strategy === Strategy.Furfi) {
-        addr = addresses['furfiStrategy'][pair][DEFAULT_CHAIN_ID]
-        abi = FurfiStrategyABI
-    }
-    return useContract(addr, abi)
+export function useWNativeContract(withSignerIfPossible?: boolean): Contract | null {
+    const { chainId } = useActiveChainId()
+    return useContract<Weth>(chainId ? WNATIVE[chainId]?.address : undefined, WETH_ABI, withSignerIfPossible)
 }
 
-export function useFreezeContract() {
-    const freezeAddr = addresses.freeze[DEFAULT_CHAIN_ID]
-    const abi = FreezeABI
-    return useContract(freezeAddr, abi)
+export function useBytes32TokenContract(tokenAddress?: string, withSignerIfPossible?: boolean): Contract | null {
+    return useContract<Erc20Bytes32>(tokenAddress, ERC20_BYTES32_ABI, withSignerIfPossible)
 }
 
-export function useReferralContract() {
-    const referralAddr = addresses.referral[DEFAULT_CHAIN_ID]
-    const abi = ReferralABI
-    return useContract(referralAddr, abi)
+export function usePairContract(pairAddress?: string, withSignerIfPossible?: boolean): IPair | null {
+    return useContract(pairAddress, IPairABI, withSignerIfPossible)
 }
 
-export function useStakingContract() {
-    const stakingAddr = addresses.stakingPool[DEFAULT_CHAIN_ID]
-    const abi = StakingABI
-    return useContract(stakingAddr, abi)
+export function useMulticallContract() {
+    const { chainId } = useActiveChainId()
+    return useContract<Multicall>(getMulticallAddress(chainId), multiCallAbi, false)
 }
+
 
