@@ -1,34 +1,60 @@
-import { createAction } from '@reduxjs/toolkit';
+import { createAction } from '@reduxjs/toolkit'
+import { ChainId } from 'src/config/constants/chains'
+import { Order } from '@gelatonetwork/limit-orders-lib'
+
+export type TransactionType =
+    | 'approve'
+    | 'swap'
+    | 'wrap'
+    | 'add-liquidity'
+    | 'remove-liquidity'
+    | 'limit-order-submission'
+    | 'limit-order-cancellation'
+    | 'limit-order-approval'
+
 
 export interface SerializableTransactionReceipt {
-    to: string;
-    from: string;
-    contractAddress: string;
-    transactionIndex: number;
-    blockHash: string;
-    transactionHash: string;
-    blockNumber: number;
-    status?: number;
+    to: string
+    from: string
+    contractAddress: string
+    transactionIndex: number
+    blockHash: string
+    transactionHash: string
+    blockNumber: number
+    status?: number
 }
 
+export enum MsgStatus {
+    MS_UNKNOWN = 0,
+    MS_WAITING_FOR_SGN_CONFIRMATIONS = 1,
+    MS_WAITING_FOR_DESTINATION_EXECUTION = 2,
+    MS_COMPLETED = 3,
+    MS_FAIL = 4,
+    MS_FALLBACK = 5,
+}
+
+
+
 export const addTransaction = createAction<{
-    chainId: number;
-    hash: string;
-    from: string;
-    approval?: { tokenAddress: string; spender: string };
-    summary?: string;
-}>('transactions/addTransaction');
-
-export const clearAllTransactions = createAction<{ chainId: number }>('transactions/clearAllTransactions');
-
+    chainId: ChainId
+    hash: string
+    from: string
+    approval?: { tokenAddress: string; spender: string }
+    claim?: { recipient: string }
+    summary?: string
+    translatableSummary?: { text: string; data?: Record<string, string | number> }
+    type?: TransactionType
+    order?: Order
+}>('transactions/addTransaction')
+export const clearAllTransactions = createAction('transactions/clearAllTransactions')
+export const clearAllChainTransactions = createAction<{ chainId: ChainId }>('transactions/clearAllChainTransactions')
 export const finalizeTransaction = createAction<{
-    chainId: number;
-    hash: string;
-    receipt: SerializableTransactionReceipt;
-}>('transactions/finalizeTransaction');
-
+    chainId: ChainId
+    hash: string
+    receipt: SerializableTransactionReceipt
+}>('transactions/finalizeTransaction')
 export const checkedTransaction = createAction<{
-    chainId: number;
-    hash: string;
-    blockNumber: number;
-}>('transactions/checkedTransaction');
+    chainId: ChainId
+    hash: string
+    blockNumber: number
+}>('transactions/checkedTransaction')
